Add RBS for type checking
This commit updates the library to use Ruby 3.2. It also adds RBS files for type checking of the library. The handling of JSON parsing was corrected in the 'remove_expired_token' method of the authentication service. Entries for '.rubocop.yml' and 'Gemfile.lock' were updated to reflect the Ruby 3.2 upgrade.

New RBS files were written for type checking and a type check was added to the '.overcommit.yml'. The required Ruby version in the 'lt-google-api.gemspec' file was updated to >= 3.2 to match the upgrade.

In addition to the above changes, files were ignored in the '.gitignore' to prevent tracking unnecessary files in the repository.

These changes were necessary to keep the library up-to-date and to leverage the benefits of type checking with RBS.