Upgrade to Ruby 3.3, update dependencies, and adjust configurations
This commit upgrades the project to Ruby 3.3. It includes updates to the `.ruby-version`, `Gemfile.lock`, and `Dockerfile` files to reflect the new Ruby version. The `bundler` dependency was bumped to 2.5 with corresponding updates in `Gemfile.lock` and `gemspec`. Various gems, including `rubocop`, `steep`, and `google-apis-drive_v3`, were updated to their latest compatible versions. Updates also include improved type definitions in RBS files and refinements to method implementations in `lt-google-api`.